# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "用户管理系统",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
...
        "/api/user/admin/delete": {
            "get": {
                "description": "DeleteUser user information by user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "DeleteUser user by user account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Useraccount",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DeleteUser successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "DeleteUser failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/admin/query": {
            "post": {
                "description": "Query user list for admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Query user list for admin",
                "parameters": [
                    {
                        "description": "queries",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modeluser.AdminUserQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/modeluser.ReturnAdminUser"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Query failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/admin/update": {
            "post": {
                "description": "admin edit user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "admin edit user information",
                "parameters": [
                    {
                        "description": "User Information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modeluser.EditUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EditUserRequest successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "EditUserRequest failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "Authenticate user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User Information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modeluser.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Common successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modeluser.ReturnUser"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Common failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/logout": {
            "get": {
                "description": "User logout",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Logout failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/query": {
            "post": {
                "description": "Get user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Query user",
                "parameters": [
                    {
                        "description": "queries",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modeluser.UserQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/modeluser.ReturnUser"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Query failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User register",
                "parameters": [
                    {
                        "description": "User Information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modeluser.AddUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success registered",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Register failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "modeluser.AdminUserQueryRequest": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "is_delete": {
                    "description": "是否删除",
                    "type": "integer"
                },
                "user_account": {
                    "description": "用户账号",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "user_role": {
                    "description": "匿名用户，普通用户，管理员，禁用用户",
                    "type": "string"
                }
            }
        },
        "modeluser.EditUserRequest": {
            "type": "object",
            "required": [
                "user_account"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "user_account": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                },
                "user_role": {
                    "type": "string"
                }
            }
        },
        "modeluser.LoginUserRequest": {
            "type": "object",
            "required": [
                "user_account",
                "user_password"
            ],
            "properties": {
                "user_account": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                },
                "user_password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 7
                }
            }
        },
        "modeluser.ReturnAdminUser": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_account": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_role": {
                    "type": "string"
                }
            }
        },
        "modeluser.ReturnUser": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_account": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "utils.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}